+++ src/borrow.rs
@@ -68,7 +68,7 @@ impl<'a, T: Component> Ref<'a, T> {
     ) -> Result<Self, MissingComponent> {
         let target = NonNull::new_unchecked(
             archetype
-                .get::<T>()
+                .get_base::<T>()
                 .ok_or_else(MissingComponent::new::<T>)?
                 .as_ptr()
                 .add(index as usize),
@@ -107,7 +107,7 @@ impl<'a, T: Component> RefMut<'a, T> {
     ) -> Result<Self, MissingComponent> {
         let target = NonNull::new_unchecked(
             archetype
-                .get::<T>()
+                .get_base::<T>()
                 .ok_or_else(MissingComponent::new::<T>)?
                 .as_ptr()
                 .add(index as usize),
@@ -187,6 +187,16 @@ impl<'a> EntityRef<'a> {
         self.archetype
             .into_iter()
             .flat_map(|arch| arch.types().iter().map(|ty| ty.id()))
+    }
+
+    /// Number of components in this entity
+    pub fn len(&self) -> usize {
+        self.archetype.map_or(0, |a| a.types().len())
+    }
+
+    /// Shorthand for `self.len() == 0`
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
     }
 }