+++ src/serialize/row.rs
@@ -1,15 +1,13 @@
-//! Convenience tools for serializing [`World`]s; requires the `serde` feature
+//! Human-friendly row-major serialization
 //!
-//! [`Component`]s are not necessarily serializable, so we cannot directly implement [`Serialize`]
-//! for [`World`]. The helper functions defined in this module allow serialization and
-//! deserialization to rely on purpose-defined traits to control the procedures explicitly.
+//! Stores each entity's components together. Preferred for data that will be read/written by
+//! humans. Less efficient than column-major serialization.
 //!
 //! This module builds on the public [`World::iter()`] and [`World::spawn_at()`] APIs, and are
-//! somewhat opinionated. For some applications, a more custom approach may be preferable.
+//! somewhat opinionated. For some applications, a custom approach may be preferable.
 //!
 //! In terms of the serde data model, we treat a [`World`] as a map of entity IDs to user-controlled
-//! maps of component IDs to data. Backwards-incompatible changes to this model are subject to the
-//! same semantic versioning stability guarantees as the hecs API.
+//! maps of component IDs to data.
 
 use core::{cell::RefCell, fmt};
 
@@ -34,12 +32,13 @@ use crate::{Component, EntityBuilder, EntityRef, World};
 /// # struct Position([f32; 3]);
 /// # #[derive(Serialize)]
 /// # struct Velocity([f32; 3]);
-/// use hecs::{*, serialize::*};
+/// use hecs::{*, serialize::row::*};
+///
+/// #[derive(Serialize, Deserialize)]
+/// enum ComponentId { Position, Velocity }
 ///
 /// // Could include references to external state for use by `serialize_entity`
 /// struct Context;
-/// #[derive(Serialize, Deserialize)]
-/// enum ComponentId { Position, Velocity }
 ///
 /// impl SerializeContext for Context {
 ///     fn serialize_entity<S>(
@@ -63,6 +62,15 @@ pub trait SerializeContext {
     fn serialize_entity<S>(&mut self, entity: EntityRef<'_>, map: &mut S) -> Result<(), S::Error>
     where
         S: SerializeMap;
+
+    /// Number of entries that [`serialize_entry`] will produce for `entity`, if known
+    ///
+    /// Defaults to `None`. Must be overridden to return `Some` to support certain serializers, such
+    /// as bincode.
+    fn component_count(&self, entity: EntityRef<'_>) -> Option<usize> {
+        let _ = entity;
+        None
+    }
 }
 
 /// If `entity` has component `T`, serialize it under `key` in `map`
@@ -106,7 +114,7 @@ impl<'a, C: SerializeContext> Serialize for SerializeComponents<'a, C> {
     {
         let mut this = self.0.borrow_mut();
         let entity = this.1.take().unwrap();
-        let mut map = serializer.serialize_map(None)?;
+        let mut map = serializer.serialize_map(this.0.component_count(entity))?;
         this.0.serialize_entity(entity, &mut map)?;
         map.end()
     }
@@ -133,12 +141,13 @@ where
 /// # struct Position([f32; 3]);
 /// # #[derive(Deserialize)]
 /// # struct Velocity([f32; 3]);
-/// use hecs::{*, serialize::*};
+/// use hecs::{*, serialize::row::*};
+///
+/// #[derive(Serialize, Deserialize)]
+/// enum ComponentId { Position, Velocity }
 ///
 /// // Could include references to external state for use by `deserialize_entity`
 /// struct Context;
-/// #[derive(Serialize, Deserialize)]
-/// enum ComponentId { Position, Velocity }
 ///
 /// impl DeserializeContext for Context {
 ///     fn deserialize_entity<'de, M>(
@@ -298,10 +307,10 @@ mod tests {
     mod helpers {
         use super::*;
         pub fn serialize<S: Serializer>(x: &World, s: S) -> Result<S::Ok, S::Error> {
-            crate::serialize::serialize(x, &mut Context, s)
+            crate::serialize::row::serialize(x, &mut Context, s)
         }
         pub fn deserialize<'de, D: Deserializer<'de>>(d: D) -> Result<World, D::Error> {
-            crate::serialize::deserialize(&mut Context, d)
+            crate::serialize::row::deserialize(&mut Context, d)
         }
     }